Track start and end line/column information for tokens

+++ src/lexing/lex_errors.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum LexError {
     UnexpectedEof,
-    InvalidToken { line: u32 },
+    InvalidToken { line: usize },
 }
