+++ src/parsing/parse_errors.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum ParsingError {
     UnexpectedEof,
-    UnexpectedToken { line: u32, message: String },
+    UnexpectedToken { line: usize, message: String },
 }
